Move provider files into provider package
